# Speed up `nth()` family