Closes https://github.com/tidyverse/dplyr/issues/6682

I'm not too worried about `vec_as_location2()` actually being slow right now. We can improve that later when we finally add `vec_slice2()` for real.

``` r
library(dplyr)

x <- 1:5

bench::mark(first(x), iterations = 100000)

# CRAN dplyr 1.1.0
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 first(x)     49.5µs   65.2µs    15182.     152KB     16.7

# This PR
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 first(x)     18.9µs   25.2µs    38844.     133KB     15.5
```

Further improvements will come from vctrs with https://github.com/r-lib/vctrs/issues/1785 and https://github.com/r-lib/vctrs/issues/1784